Infoscreen for guild room (#565)
* wip: info screen

* Add files to branch after big fuckup

* add fetcher

* debug

* Hello

* hups

* Add debug message

* fix process.env.* functioning

* Fix issue where events next year show as the first events on the list :D

* something

* Change PUBLIC_**_URL to NEXT_PUBLIC...

* More infonäyttö things

- Fixed scrolling between pages
- Store state at parent component
- Misc things

* fix: Formatted infonäyttö files

* fix: Changed rest of PUBLIC_ env variable references to NEXT_PUBLIC_

* Change PUBLIC_**_URL to NEXT_PUBLIC_

* fix: Align stop names on HSL schedule with grid

* Refactor folder structure for infoscreen components

* remove hsl page

* format

* add kanttiinit functionality

* Add kanttiinit stuff

* feat: Update Kanttiinit implementation

- Now it works :D
- Created separate API endpoints for accessing Kanttiinit API
- Changed types for Kanttiinit
- Updated almost all files that touch Kanttiinit on infoscreen

* feat: Added basic styling to Kanttiinit on infoscreen

* Something

* Make infoscreen fullscreen capable

* fix some typecheck issues

* format and lint

* Fix Raide-Jokeri writing form

* fix: Fixed typecheck errors in Infoscreen

- Full screen button now only works on new browsers (Safari 2023+, most others 2020+)
- Removed unused Event component
- Better error response for Kanttiinit api handler

* Removed Enter Fullscreen button from InfoScreen

* fix: Fixed formatting

* Revert global env variable changes

* Remove non-relevant files, fixed some issues

* Removed commented code, console.logs etc

* Fixed Header logo on Infoscreen, 2025 payload update to make tests pass, removed unused json-stable-stringify

* Fixed formatting and some warnings

* Moved dependency to correct package.json

* Added one newline to fix formatting :)

* Fix several warnings

* Move types, small refactoring, etc

* Revert infoscreen layout/page structure change

- It seems that it is mandatory to have both layout and page files

* Refactored infonaytto contents file

* Convert HSL & Kanttiinit to Server Components on infoscreen

- Event page is currently not working

* chore: cleanup for hsl stuff mostly

* chore: Add caching for Kanttiinit API, fix arrivalTimeUnix parameter name

* feat: Initial event display support for infoscreen

* feat: Refactor Infoscreen events page to use global event components

- Added new EventCardCompact component

* feat: Add localization to infoscreen, minor layout fixes

* feat: Fetch localized Kantiinit menus, change infoscreen path

- Infoscreen is now /infoscreen instead of /infonaytto/naytto
- Kanttiinit menus are now fetched in Finnish and English depending on locale

* refactor: Update infoscreen component structure

- Renamed `kanttiinit-combined` to just `kanttiinit`
- Combined kanttiinit helper functions to `kanttiinit/fetcher.ts`
- Moved locale handling to kanttiinit React component
- Combined `hsl-schedules-combined` and `hsl-schedule` to `hsl-schedules`
- Update event grouping week logic on infoscreen
- Revert the change to NEXT_PUBLIC_ILMOMASIINA_URL
- Revert moving of `globals.css`

* Fix: uncomment HSL and Kanttiinit screens in InfoScreenContents

---------
