+++ .env.example
@@ -16,8 +16,8 @@ PAYLOAD_PUBLIC_LOCAL_DEVELOPMENT=true
 
 NEXT_REVALIDATION_KEY="veryprivatekey"
 
-# Digitransit API key for HSL traffic data for info screen www.digitransit.fi/en/developers/api-registration/
-DIGITRANSIT_SUBSCRIPTION_KEY="very secret stuff"
+# Digitransit API key for hsl traffic data for info screen www.digitransit.fi/en/developers/api-registration/
+DIGITRANSIT_SUBSCRIPTION_KEY="Replace me"
 
 PUBLIC_FRONTEND_URL="http://localhost:3000"
 PUBLIC_SERVER_URL="http://localhost:3001"
