+++ apps/web/package.json
@@ -26,6 +26,7 @@
     "react": "catalog:react19",
     "react-big-calendar": "^1.17.0",
     "react-dom": "catalog:react19",
+    "react-live-clock": "^6.1.22",
     "react-markdown": "^9.0.1",
     "remark": "^15.0.1",
     "remark-gfm": "^4.0.0",
