+++ apps/web/src/app/[locale]/(infoscreen)/infoscreen/page.tsx
@@ -1,0 +1,16 @@
+import { HSLcombinedSchedule } from "../../../../components/infoscreen/hsl-schedules";
+import { KanttiinitCombined } from "../../../../components/infoscreen/kanttiinit";
+import InfoScreenSwitcher from "../../../../components/infoscreen/infoscreen-switcher/index";
+import EventListInfoscreen from "../../../../components/infoscreen/events-list";
+
+export const dynamic = "force-dynamic";
+
+export default function InfoScreenContents() {
+  return (
+    <InfoScreenSwitcher>
+      <HSLcombinedSchedule />
+      <KanttiinitCombined />
+      <EventListInfoscreen />
+    </InfoScreenSwitcher>
+  );
+}
