+++ apps/web/src/app/[locale]/(main)/events/[...ilmomasiinaPath]/page.tsx
