+++ apps/web/src/locales/fi.ts
@@ -41,6 +41,10 @@ const fi = {
   "generic.Version": "Versio",
   "heading.Main navigation": "Päävalikko",
   "heading.Upcoming events": "Tulevat tapahtumat",
+  "infoscreen.Ruokalistat": "Ruokalistat",
+  "infoscreen.Bussit": "Bussit",
+  "infoscreen.Raide-Jokeri": "Raide-Jokeri",
+  "infoscreen.Metro": "Metro",
   "invoicegenerator.Invoicer name": "Laskuttajan nimi",
   "invoicegenerator.Invoicer email": "Laskuttajan sähköpostiosoite",
   "invoicegenerator.Phone number": "Puhelinnumero",
@@ -128,8 +132,10 @@ const fi = {
   "ilmomasiina.status.Ei ilmoittautuneita vielä": "Ei ilmoittautuneita vielä.",
   "ilmomasiina.status.Ilmoittautuminen alkaa":
     "Ilmoittautuminen alkaa {startDate}",
+  "ilmomasiina.status.Ilmo alkaa": "Ilmo alkaa {startDate}",
   "ilmomasiina.status.Ilmoittautuminen auki":
     "Ilmoittautuminen auki {endDate} asti",
+  "ilmomasiina.status.Ilmo auki": "Ilmo auki {endDate} asti",
   "ilmomasiina.status.Ilmoittautuminen on päättynyt":
     "Ilmoittautuminen on päättynyt",
   "ilmomasiina.status.Ilmoittautumistiedot eivät ole julkisia":
